% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_arguments.R
\name{setDefault}
\alias{setDefault}
\alias{check_and_set_Default}
\title{Sets variable to default value}
\usage{
setDefault(var, defaults, envir = parent.frame(), verbose = TRUE)

check_and_set_Default(var, defaults = NULL, envir = parent.frame(),
  inherits = FALSE, verbose = TRUE)
}
\arguments{
\item{var}{Name of object to set to default value.}

\item{envir}{Environment in which to set the value.}

\item{default}{Named list of default values.  The name of one (and only one) element must match the provided string in \code{var}.}
}
\details{
\code{check_and_set_Default} First checks to see if a variable has been set to a value. If not, it calls \code{setDefaul} to set to default value.
}
\seealso{
assign
}
\author{
Joshua Keller
}
